Move Bluetooth management into a separate class
Here's a suggestion for separating your Activity from the
management of the Bluetooth state. By using a BehaviorSubject, the last
known state is cached and you don't have to send yourself the state in
onResume. I'm not too familiar with RxJava, so casting the subject to an
observable may be necessary or right - caveat emptor.